#185 Added claim checker after function is called
There are times where (i believe) that the code for checking the claim was done before the cloud function actually completes (it takes 1 sec to complete I think) and so I moved it to a then function.

Added a debugPrintMeta to print some app metadata (is this how you use meta?)

Updated firestore rules.